Reduce required election time
The election period is a stressful time for candidates. The week long election period increases the amount of campaigning required by candidates and the period with which they must be anxious about the election without much gain. Three days is adequate time for members to discern the candidates they wish to vote for and to vote. To further ensure everyone who wishes to vote hears about the election, clause 8.4.4  stipulates that the election should be properly advertised.